requirements.txt
altair==4.2.0
altair_viewer==0.4.0
branca==0.5.0
databases==0.5.5
fastapi==0.78.0
folium==0.12.1.post1
geopy==2.2.0
Jinja2==2.11.3
leafmap==0.9.3
matplotlib==3.5.1
mysql_connector_repackaged==0.3.1
nltk==3.7
numpy==1.22.2
pandas==1.4.2
Pillow==9.1.1
psycopg2==2.9.3
pydantic==1.9.0
pyodbc==4.0.32
pyotp==2.6.0
pyzbar==0.1.9
requests==2.27.1
seaborn==0.11.2
seleniumbase==3.2.8
setuptools==62.1.0
sql==2022.4.0
SQLAlchemy==1.4.32
streamlit_option_menu==0.3.2
